When DSS writes messages, use message types in output file name
To make it possible to know from the file name what type of message it contains.
This is useful for external tools to know what type of message a file
contains without parsing the JSON; just looking at the file name is
significantly cheaper if the JSON would only be parsed for this.

We have this case in multi-processing DSS, where the scheduler only
needs to know the type of messages to forward them correctly

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/abstraction-with-block-summaries@48210 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c