Refactor start script of CPAchecker
Previously the code for handling stuff like the heap size
was spread over the whole script: a default at the top,
some logic for handling --benchmark, and some code further down
for logging and determining the final value.

Now we just set a flag for benchmark mode,
and can then have a single place in the code
for each thing to handle (like stack size, heap size, etc.).
Only the parsing of the option still needs to be separate.