Only compute correctness witness internally if needed
Part of #1175. The internal computation of the Witness instance
can be quite slow, and so far we did this always, even if we would
ignore the result and not write it to disk.
The only case where the computation was skipped was if all output files
of the ARGCPA were disabled, but this does not help for cases like
SV-COMP where we want other witnesses to be written.
With this commit we only compute the Witness instance if necessary.

For the case mentioned in #1175 the writing of output files is now much
faster, though it still takes more than 20s for me.