Mark analysis as terminated in ParallelAlgorithm in case of exception
d11eb920 implemented skipping of statistics output for analyses
that are still running after the ParallelAlgorithm terminates,
but the flag for termination was set only if the analysis terminates
regularly. Now we also set it in case of an exception such that we get
statistics instead of the wrong message
"Cannot print statistics for ... because it is still running."

We could probably also get rid of this flag completely and check whether
the respective future is finished, but this is a bigger refactoring.
Part of #326.