Do not share reached set from ParallelAlgorithm redundantly
If the inner algorithm is a ReachedSetUpdateListener
and the user specified the ::supply-reached-refinable annotation,
the reached set of the analysis would be shared twice,
via both ReachedSetUpdateListener and the ::supply-reached-refinable
implementation.
This is likely undesired, and in any case the whole reached-set sharing
is much easier to understand if we only use one sharing mechanism per
analysis.
So now ReachedSetUpdateListener is only used if
::supply-reached-refinable is not given.

This should have no effect in practice because we use
::supply-reached-refinable only for the InvariantsCPA configs,
which do not use CEGARAlgorithm and thus do not have a
ReachedSetUpdateListener.

There is another case of redundant reached-set sharing,
but this is more difficult to remove, so just clarify the comment.