Implemented refinement in ParallelAlgorithm without reached-set sharing
So far one could use ::supply-reached-refine to enable a special mode of
the ParallelAlgorithm where it executes the analysis in an iterative
manner. But this also turned on sharing of the reached set with other
anlyses. In cases where the analyses are not suited for this the sharing
can be unsound, and it is inefficient if nobody consumes it.
So now we not only provide ::supply-reached and
::supply-reached-refinable but also ::refinable.
Internally we already had two boolean flags of which only three possible
value combinations were used, so we can just continue to use these now
orthogonal two flags.

This improves the situation of the data-flow analysis that is discussed
in #1156.