Disable witness tab in HTML reports by default
Generating the witness can take a long time (cf. #1175),
and for the report we would even regenerate it
instead of taking the witness that we compute for writing it to disk
(cf. #1268).
As long as that is not fixed, let's disable the witness tab
such that the performance of CPAchecker is not negatively affected by
it. Users can still enable it with a configuration option.