Catch CPAExceptions during analysis creation in ParallelAlgorithm
Since 56e9a19d the ParallelAlgorithm would propagate all exceptions
that occur during instantiation of the analyses immediately.
This means that if one analysis cannot be instantiated,
we give up completely.
However, there are expected cases where an analysis cannot be created
but we want to continue with other analyses,
such as if the loop information is missing.
So let's catch and log these exceptions instead.

This does not completely revert 56e9a19d because we still propagate
other exceptions such as InvalidConfigurationExceptions.
We want to do this because of #318.