[Violation witness validation v2.0] When entering the error state of the automaton copy the target information of the property being violated
When validating witnesses we set the option `cpa.automaton.treatErrorsAsTargets` to `false` in order to match only the target states of the witness as errors and not the target states of the specification. In order to get the correct specification violation in the output we then copy the information from the specification into the witness target state. This was currently wrong for the first edge which enters the violation. This commit addresses that error.

Closes: https://gitlab.com/sosy-lab/software/cpachecker/-/issues/1250