Fix termination of CPAchecker in case non-daemon threads are running
In a parallel portfolio it may happen that one analysis thread
does not react fast enough to shutdown requests and
is still running after everything finishes and the statistics are printed.
In this case we do not want to wait for it, but force JVM termination.
This commit implements this by calling System.exit() at the end of the
main method instead of just letting the main method finish.
And we add some warning with information about which threads did not
terminate.