Update test-sets using the latest information from the benchdefs repository for CPAchecker
There were many tasks added in the last couple of days due to SV-COMP25.
This commit updates our tests in order to also include them.
Some of the tests have special filtering rules which would also need to
be updated, but this will require some specialized knowledge about the
analysis and what needs to be done.