Turn assertions in RankingRelationBuilder into regular checks
There are some unsupported cases in ranking functions,
and some where checked only with an assert statement.
This had two effects: Whenever asserts were disabled
(e.g., in --benchmark runs), the code would just continue
and assume that the unsupported term was a memory access.
This cannot be intended.
Furthermore, with asserts enabled the code would simply crash.

Issues #937 and #1302 are cases where these asserts trigger.

This change turns the asserts into regular checks
and throws a RankingRelationException instead.
This exception is actually caught by the call chain,
so CPAchecker will not terminate with an error
but continue looking for other termination arguments,
eventually leading to either a real result, giving up, or a timeout.
(For some other cases of unsupported terms this was already done.)
I am not sure whether this is useful or mostly a waste of resources
if we continue, this part is open for discussion.
But at least we should produce a proper error message instead of a
crash if we decide to change this.