Add CI job that automatically updates dependencies and creates an MR
This was previously done by BuildBot, which then executed the CI checks
and committed directly to trunk.
With GitLab we can now improve this.
We check for updates and commit and push any resulting changes to a
branch. Then we open an MR and let the usual CI pipeline check whether
everything still works.
If yes, a developer can approve and the MR gets merged.
If there are manual fixes required, these can be pushed to the MR.