Introduce ghost edge in DSS decomposition just before the block end, not as new block end
This solves two issues:
(1) If the block end is a branching node with two outgoing edges,
adding a new ghost edge would introduce a thirds edge. Not every algorithm in CPAchecker likes this,
because they assume well-formed CFAs.
(2) If the block end is a function-exit node, it not only has a return edge,
but also a leaving blank edge. This also violates the well-formedness guarantees that many
analyses assume about the CFA.