Refactoring: control-flow simplification to avoid continue (I)
Instead of having 3 cases with one containing just "continue"
and one being an "else" branch,
it is easier to have just two cases,
because one has to remember less other code
to understand under which conditions the last branch is used.

This code was flagged by the RedundantControlFlow check of Error Prone.

Part of #1307.