Refactoring: control-flow simplification to avoid continue (III)
Instead of using "continue" to skip the next branch
using an "if-else-chain" would be better.
But in the cases here the code in the branches is even the same,
so we can just merge everything into one condition.

This code was flagged by the RedundantControlFlow check of Error Prone.

Part of #1307.