Cleanup statistics code in CachingPathFormulaManager
c39c43f1 attempted to make the statistics thread safe,
but this is completely useless in a class that is not thread safe itself
(because it uses unsynchronized HashMaps as caches).
It could even mislead people to think that the class is thread safe.

So it is better to have standard code here, and we want to get rid of
ThreadSafeTimerContainer (which is not thread safe) anyway.