Remove usage of ThreadSafeTimerContainer from PredicateCPA
This was an attempt to make the PredicateCPA thread safe,
but this did not work out according to
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/70298e2de1b4d1a258e57e2ba34c07636ec6e72f#note_647467995
and will likely never work because solvers are not thread safe.
Furthermore, ThreadSafeTimerContainer is not thread safe
and we want to get rid of it because it usage leads to wrong
assumptions.

This effectively reverts most of 3ff98c90.