Use a single instance of the precision adjustment in the PredicateCPA
The use of separate instances of operators was
intended to allow multi-threaded usage of the PredicateCPA.
But this did not work out according to
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/70298e2de1b4d1a258e57e2ba34c07636ec6e72f#note_647467995
and will likely never work because solvers are not thread safe.

This change moves the statistics of PredicatePrecisionAdjustment back
into this class, which is better for code coherency.
Furthermore, they were previously even stored inconsistently.

This reverts parts of 2709a5c2.