Use a single instance of PredicateAbstractionManager in the PredicateCPA
The use of separate instances of PredicateAbstractionManager was
intended to allow multi-threaded usage of the PredicateCPA.
But this did not work out according to
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/70298e2de1b4d1a258e57e2ba34c07636ec6e72f#note_647467995
and will likely never work because solvers are not thread safe.

The change to PredicateAbstractionManager was actually problematic,
because this class uses several caches and it was not clear whether
these caches actually worked as intended with several different
instances of PredicateAbstractionManager (each with its own cache).
So we want to fix this.

Furthermore, this allows us to move the statistics and options of
PredicateAbstractionManager back into the class,
which improves the encapsulation, requires callers of the constructor to
know less about its implementation details, makes statistics in
PredicateAbstractionManager more consistent (some statistics were stored
directly in the class, some statistics were stored in a separate class),
and requires less code overall.

This reverts most of 89bfef3e, 65125f36, and adc5330a.