Fix statistics of CachingPathFormulaManager
The number of cache hits and misses was counted differently
(makeEmptyPathFormulaWithContextFrom incremented the number of hits
but not the number of misses), so the percentage was wrong.
The time spent for PathFormula creation was redundant
because it is already computed and printed by other code.
Now this is cleaned up, and there is a strict and easy to see
1:1 correspondence for the counters for cache hits and misses.

Additionally, we move the statistics to a better place in the output,
because there are no time measurements left.