Activate RedundantControlFlow checker of Error Prone
There is only one remaining warning left,
and it is unclear whether this can be fixed soon,
so it is better to suppress it and enable the check.

There is also the related DuplicateBranches check,
which we cannot enable right now, but we can make some preparations for
it (eliminating redundant code and improving comments).

Closes #1307