Better implementation of per-analysis CPU time in ParallelAlgorithm
Since fa6b83e6 the measurement of the per-analysis CPU-time statistics
was implemented using the ThreadCpuTimeLimit class,
whose main purpose is to be a resource *limit*, not measurement.

This had several disadvantages:
- The API of ResourceLimitChecker had to be adjusted to allow access to
  its internal limit implementations.
- ResourceLimitChecker and ThreadCpuTimeLimit had to take do some
  specific steps to guarantee somewhat precise measurements,
  which would not be necessary for their main purpose.
- Even so, the measurement was not as precise as it could be
  because the measurement was done in a different thread
  that did not know when exactly the measured thread ended.
- CPU time was only measured for those analyses that a per-thread
  CPU-time limit configured, not for unlimited analyses.
- A race condition could lead to crashes (#1296).

This commit separates the measurement from the limit implementation
and fixes all these disadvantages, while at the same time needing less
code and making the implementation of the limit much more
understandable: We simply need to read the CPU-time usage of the
analysis thread directly before the thread ends.

Fixes #1296.