Refactor resource limits to have explicit start() method
So far, ResourceLimit implementations tracked the time starting from
when they were created. This did not work for ThreadCpuTimeLimit and it
was changed to have a special method for indicating the start in !187.
But this seems actually to be a good idea in general,
and it is better to be consistent for all limits.
So now, creating a ResourceLimit instance does nothing,
and start() needs to be called at the point where the limit should start.
ResourceLimitChecker does that automatically.
The difference to before is that the time between instance creation
and start() is no longer counted for the limit.
But for all users except for ParallelAlgorithm this seems to be
irrelevant because they call start() right after creating the limits.
And for ParallelAlgorithm it only makes other limits consistent with how
the ThreadCpuTimeLimit already behaves.

Fixes #1280.