Use name of analysis for log messages in ParallelAlgorithm
When we create sub-analyses, we make the logger append an identifier for
the sub-analysis to each log message. The algorithms for strategy
selection and sequential portfolio use the name of the config file for
this, but the ParallelAlgorithm so far used just a number.
This makes it hard to understand the log output.
Now we also use the name of the config file here,
which fixes this but makes the suffix of the log messages in complex
configurations even longer.