Cleanup compareTo implementations of some test classes
If we crash, it does not matter (especially in tests, but even outside)
if we crash with an UnsupportedOperationException (especially one without
a meaningful message) or a ClassCastException.
So we can simplify the code to just blindly cast.

And then the methods nicely match our recommendations,
not comments necessary to silence CI.

Only the check script needs to be extended to accept Float.compare
like it already accepts Double.compare.