Floatingpoint: Remove casting in AssumptionToEdgeAllocator.getValueLiteral and throw an exception if the value and type don't match
The casting code was added to address issues in ConcreteErrorPathAllocator, but the fix was always incomplete. As it is now covering up some crashes related to https://gitlab.com/sosy-lab/software/cpachecker/-/merge_requests/228 we're removing it again. For the issues in ConcreteErrorPathAllocator a separate issue was already opened here: https://gitlab.com/sosy-lab/software/cpachecker/-/issues/1245