Refactoring: use pattern-matching instanceof (NEG)
This introduces pattern-matching instanceof for cases that
- have a negated instanceof check,
- where no method call exists in the expression, and
- where the value is not assigned to a variable after the cast.

Because the value is never assigned to a variable in the previous code
but now it is, a name for the new variable needs to be found.
Error Prone creates a variable name based on the class in the instanceof
check, and I guess this is the best that we can do automatically.
In some cases this does reduce the readability of the code a little bit,
for example if "(CIdExpression) lhs" is replaced with "cIdExpression",
but in many cases the chosen variable name actually fits the existing
code, because we have plenty of variables that are named after their type.

Commit created with Google Error Prone.