correctly assign start_routine arguments when creating threads
this commit also fixes an issue where incorrect amounts of thread simulations were created in the sequentialization, e.g. when the pthread_create call is inside a function which is called multiple times. the input CFA of CPAchecker contains only one pthread_create edge, but it is called multiple times, which is handled correctly now thanks to the call context sensitive ThreadEdges in the ThreadCFA.