Refactoring: avoid fall through by moving code into branch
Here we can unify two cases if we put the extra code from the first case
into an additional branch.
The additional check for pOpType is not perfect,
but getting rid of all fall throughs is worth it.
And one could argue that the new code is even easier to understand
due to being more explicit and the handling of op1 being close to the
same handling of op2 now.