Refactoring: use switch expression and avoid partially used builder
The code here always created a builder, but only used it in some cases.
In other cases it circumvented its use with a direct return.
This could be confusing and lead to bugs
(addition of a value to the builder but then ignoring it),
so it is better to only create the builder when necessary
(which actually is only a single case).