Enable Error Prone's PatternMatchingInstanceOf
All suggestions of it have been applied in the previous commits.

Error Prone has the option
-XepOpt:ConstantExpressions:ConsiderAllMethodsPure=true
that makes PatternMatchingInstanceOf apply to more cases
(even if there is a method call in the instanceof expression).
We consider this to be fine and would like to enable it,
but ConsiderAllMethodsPure also changes the behavior of other
Error Prone checks for which it does not make sense.
So we have to live without it.