Use Java 21 in our container images
Even though Java 17 also works,
we have seen that Java 21 is often faster.

The Ubuntu package openjdk-21-jre-headless has some dependencies
marked only as "recommended" that were hard dependencies for 17.
We do not install all recommended dependencies because that includes
sound and printing support, which would be useless.
For the same reason we do not use the non-headless package.
But font rendering is now also a recommended dependency,
so we manually need to include those such that our SVG export works.