Refactor handling of unary expressions in SMGCPAValueVisitor
Remove common code of branches and make the actual effect of the loop
more visible. This also avoids potential bugs because now it is
guaranteed that every branch produces a new value.
Also avoids use of continue and uses a switch expression.
As a side effect, SpotBugs no longer stumbles about the continue in a
switch.